Ensure matching exact binary name.
This avoids llvm-vim-plugins which was dropped and thus only detects older
version. A full filename match was the approach one would expect from the
get go, but it would appear it was done to avoid having to introduce the
package/binary .ini syntax that ended up being a necessity. As such there
is no further reason to avoid a full match.